Changed constraints and added all example data
README.md file has the updated schema changes and reflects that the
example data has been finished (side note holy shit was this time-
consuming!!!). The tables.sql file has the new
schema, with the removal of GroupID from the MESSAGES table. It
was removed because we can use a stored procedure to query all of
the users in a group and send a message to each using a cursor.
- Added a DateCreated to the Users table
- All timestamp values are now precision 3 instead of 6
- Removed GroupID from Messages table
- All example data for tables is included