Finished initiateFriendship(). Fixed establishFriendship()
initiateFriendship() method finished. Added header to the output that explains
what data is being shown. Also changed the input to retrieve an email
address instead of firstname/lastname because the email field is unique
within the Users table.

Also fixed a typo where establishFriendship was using the userEmail
variable for both the user email AND the friend email.