Testing searchForUser(). Needs Finishing.
Changed the searchForUser() method to not return duplicate results. The
problem with this method though is it is NOT SAFE because it is not
using the placeholders of the PreparedStatement object, and instead
I am just putting in raw String values.

The problem before was duplicate results. I wonder if there is a way
to marry the fix with a valid and safe preparedStatement placeholder.